Refactor mqtt connection into custom effect
