+++ client/src/location/mqttConnection.ts
@@ -1,11 +1,91 @@
+import { useState, useEffect } from 'react';
+import { default as UbiMqtt } from 'ubimqtt';
 import Deserializer, {
   BabylonBeacon,
   mqttMessageToBabylon,
+  BeaconGeoLocation,
+  MqttMessage,
+  mqttMessageToGeo,
 } from './mqttDeserialize';
+import { currentEnv } from '../common/environment';
+import { DEFAULT_TOPIC } from '../location/urlPromptContainer';
 
 const MOCK_MESSAGE_INTERVAL = 2000;
 
 const ROOM_HEIGHT_METERS = 3.8;
+
+export const refreshBeacons = (
+  parsed: MqttMessage[],
+  bluetoothName: string | null,
+  lastKnownPosition: BeaconGeoLocation | null
+) => {
+  const geoBeacons = parsed.map(i => mqttMessageToGeo(i));
+
+  const ourBeacon = geoBeacons.find(
+    beacon => beacon.beaconId === bluetoothName
+  );
+
+  return {
+    beacons: geoBeacons,
+    lastKnownPosition: ourBeacon !== undefined ? ourBeacon : lastKnownPosition,
+  };
+};
+
+export const useUbiMqtt = (
+  host: string,
+  bluetoothName: string | null,
+  topic?: string
+) => {
+  const parser = new Deserializer();
+
+  const [beacons, setBeacons] = useState<BeaconGeoLocation[]>([]);
+  const [
+    lastKnownPosition,
+    setLastKnownPosition,
+  ] = useState<null | BeaconGeoLocation>(null);
+
+  useEffect(() => {
+    if (!currentEnv.MAPBOX_TOKEN) {
+      console.error('mapbox api token missing, falling back to raster maps...');
+    }
+
+    const ubiClient = new UbiMqtt(host, { silent: true });
+    ubiClient.connect((error: any) => {
+      if (error) {
+        console.error('error connecting to ubi mqtt', error);
+      } else {
+        console.log('connected to', host);
+        ubiClient.subscribe(
+          topic || DEFAULT_TOPIC,
+          null,
+          (connectedTopic: string, msg: string) => {
+            const nextBeacons = refreshBeacons(
+              parser.deserializeMessage(msg),
+              bluetoothName,
+              lastKnownPosition
+            );
+
+            setBeacons(nextBeacons.beacons);
+            setLastKnownPosition(nextBeacons.lastKnownPosition);
+          },
+          (err: any) => {
+            if (err) {
+              console.error('error during sub', err);
+            }
+          }
+        );
+      }
+    });
+
+    return () => {
+      ubiClient.forceDisconnect(() => {
+        console.log('disconnected from ubimqtt');
+      });
+    };
+  }, []);
+
+  return { beacons, lastKnownPosition };
+};
 
 export class FakeMqttGenerator {
   intervalRef: number;
